Fixed access groups in sql scripts, added utility file for sql queries and database methods
